requirements.txt
av==14.4.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
faster-whisper==1.1.1
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
onnxruntime==1.22.0
packaging==25.0
pillow==11.0.0
protobuf==6.31.1
PyAutoGUI==0.9.54
pycparser==2.22
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pystray==0.19.5
pytweening==1.2.0
PyYAML==6.0.2
requests==2.32.4
scipy==1.15.3
six==1.17.0
sounddevice==0.5.2
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.5.0
